[index][util] Move tuples metafunctions from index to util.
