+++ include/boost/geometry/util/tuples.hpp
@@ -2,21 +2,21 @@
 //
 // Copyright (c) 2011-2013 Adam Wulkiewicz, Lodz, Poland.
 //
+// This file was modified by Oracle on 2019.
+// Modifications copyright (c) 2019 Oracle and/or its affiliates.
+// Contributed and/or modified by Adam Wulkiewicz, on behalf of Oracle
+//
 // Use, modification and distribution is subject to the Boost Software License,
 // Version 1.0. (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef BOOST_GEOMETRY_INDEX_DETAIL_TUPLES_HPP
-#define BOOST_GEOMETRY_INDEX_DETAIL_TUPLES_HPP
+#ifndef BOOST_GEOMETRY_UTIL_TUPLES_HPP
+#define BOOST_GEOMETRY_UTIL_TUPLES_HPP
 
 #include <boost/tuple/tuple.hpp>
 #include <boost/type_traits/is_same.hpp>
 
-// TODO move this to index/tuples and separate algorithms
-
-namespace boost { namespace geometry { namespace index { namespace detail {
-
-namespace tuples {
+namespace boost { namespace geometry { namespace tuples {
 
 // find_index
 
@@ -197,8 +197,6 @@ struct push_back<Tuple, T, N, N>
     }
 };
 
-} // namespace tuples
+}}} // namespace boost::geometry::tuples
 
-}}}} // namespace boost::geometry::index::detail
-
-#endif // BOOST_GEOMETRY_INDEX_DETAIL_RTREE_TAGS_HPP
+#endif // BOOST_GEOMETRY_UTIL_TUPLES_HPP
